+++ frontend/src/state/features/preferences.ts
@@ -1,39 +1,5 @@
-import { Avatar } from "@/lib/avatar";
+import { Avatar, AvatarConfig } from "@/lib/avatar";
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
-
-export type AvatarConfig = {
-	hairStyle:
-		| "plain"
-		| "wavy"
-		| "shortCurls"
-		| "parting"
-		| "spiky"
-		| "roundBob"
-		| "longCurls"
-		| "buns"
-		| "bangs"
-		| "fluffy"
-		| "flatTop"
-		| "shaggy";
-	hairColor: "000000" | "1d5dff" | "ff543d" | "fff500" | "ffffff";
-	mood:
-		| "happy"
-		| "sad"
-		| "angry"
-		| "neutral"
-		| "superHappy"
-		| "hopeful"
-		| "confused";
-	skinColor: "8d5524" | "c26450" | "e6b087" | "ffd6c0" | "ffe4d3";
-	backgroundColor:
-		| "e02929"
-		| "e08529"
-		| "e0da29"
-		| "5de029"
-		| "29e0d4"
-		| "9129e0"
-		| "e029ce";
-};
 
 export interface PreferencesState {
 	volume: number;
